requirements.txt
Flask==1.1.1
click==7.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
line-bot-sdk==1.16.0
gunicorn==20.0.4
